added driver.ts file
