+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "start": "npx ts-node src/stringCalculator.ts"
+    "start": "npx ts-node src/driver.ts"
   },
   "keywords": [],
   "author": "",
