+++ src/stringCalculator.ts
@@ -1,23 +1,3 @@
-import * as readline from "readline";
-const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-  
-try{
-rl.question("Enter Strings ", (input) => {
-  console.log(`You entered: ${input}`);
-  const result = add(input); 
-  console.log(`The result is: ${result}`);
-      rl.close(); 
-});
-}catch(err)
-{
-    console.log("Error is:"+err);
-}
-
-
-
 export function add(numbers: string): number {
     try{
     if (numbers === "") return 0;
